+++ system/dev/bus/acpi/pci.h
@@ -6,5 +6,5 @@
 
 #include <zircon/syscalls/pci.h>
 
-zx_status_t get_pci_init_arg(zx_pci_init_arg_t** arg, uint32_t* size);
+zx_status_t get_pci_init_arg(zx_pci_init_arg_t** arg, size_t* size);
 zx_status_t pci_report_current_resources(zx_handle_t root_resource_handle);
